Extract MCP tools into their own files.
