+++ src/cli.ts
@@ -5,7 +5,7 @@ import { config } from "dotenv";
 import { resolve } from "path";
 import { getServerConfig } from "./config.js";
 import { startHttpServer } from "./server.js";
-import { createServer } from "./mcp.js";
+import { createServer } from "./mcp/index.js";
 
 // Load .env from the current working directory
 config({ path: resolve(process.cwd(), ".env") });
