+++ src/index.ts
@@ -1,5 +1,5 @@
 // Re-export the server and its types
-export { createServer } from "./mcp.js";
+export { createServer } from "./mcp/index.js";
 export type { SimplifiedDesign } from "./services/simplify-node-response.js";
 export type { FigmaService } from "./services/figma.js";
 export { getServerConfig } from "./config.js";
