+++ src/mcp/tools/index.ts
@@ -1,0 +1,2 @@
+export { createGetFigmaDataTool } from "./get-figma-data-tool.js";
+export { createDownloadFigmaImagesTool } from "./download-figma-images-tool.js";
